IllegalArgumentException added to deleteCategory in DishCategoryDomainService.  Logs added

+++ domain-tier/src/main/java/com/outdoor/foodcalc/domain/service/dish/DishCategoryDomainService.java
@@ -5,6 +5,7 @@ import com.outdoor.foodcalc.domain.exception.NotFoundException;
 import com.outdoor.foodcalc.domain.model.dish.DishCategory;
 
 import com.outdoor.foodcalc.domain.repository.dish.IDishCategoryRepo;
+import com.outdoor.foodcalc.domain.repository.dish.IDishRepo;
 import com.outdoor.foodcalc.domain.service.product.ProductCategoryDomainService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -26,9 +27,12 @@ public class DishCategoryDomainService {
 
     private final IDishCategoryRepo categoryRepo;
 
+    private final IDishRepo dishRepo;
+
     @Autowired
-    public DishCategoryDomainService(IDishCategoryRepo categoryRepo) {
+    public DishCategoryDomainService(IDishCategoryRepo categoryRepo, IDishRepo dishRepo) {
         this.categoryRepo = categoryRepo;
+        this.dishRepo = dishRepo;
     }
 
     /**
@@ -70,9 +74,11 @@ public class DishCategoryDomainService {
      */
     public void updateCategory(DishCategory category) {
         if (!categoryRepo.exist(category.getCategoryId())) {
+            LOG.error("Dish category with id={} doesn't exist", category.getCategoryId());
             throw new NotFoundException("Dish category with id=" + category.getCategoryId() + " doesn't exist");
         }
         if(!categoryRepo.updateCategory(category)) {
+            LOG.error("Failed to update dish category with id=" + category.getCategoryId());
             throw new FoodcalcDomainException("Failed to update dish category with id=" + category.getCategoryId());
         }
     }
@@ -86,10 +92,17 @@ public class DishCategoryDomainService {
      */
     public void deleteCategory(long id) {
         if (!categoryRepo.exist(id)) {
+            LOG.error("Dish category with id={} doesn't exist", id);
             throw new NotFoundException("Dish category with id=" + id + " doesn't exist");
         }
-        if(!categoryRepo.deleteCategory(id)) {
-            throw new FoodcalcDomainException("Failed to delete dish category with id=" + id);
-        }
+        if (dishRepo.countDishesInCategory(id) == 0) {
+            if (!categoryRepo.deleteCategory(id)) {
+                LOG.error("Failed to delete dish category with id=" + id);
+                throw new FoodcalcDomainException("Failed to delete dish category with id=" + id);
+            }
+        } else {
+            LOG.error("Dish category with id = {} is not empty", id);
+            throw new IllegalArgumentException("Dish category with id=" + id + " is not empty");
+            }
     }
 }
